+++ ts/features/services/details/saga/index.ts
@@ -1,7 +1,7 @@
 import { SagaIterator } from "redux-saga";
 import { takeEvery, takeLatest } from "typed-redux-saga/macro";
 import { BackendClient } from "../../../../api/backend";
-import { ServicesClient } from "../../common/api/client";
+import { ServicesClient } from "../../common/api/servicesClient";
 import {
   loadServicePreference,
   upsertServicePreference
