+++ ts/features/services/home/saga/__tests__/handleGetFeaturedInstitutions.test.ts
@@ -5,7 +5,7 @@ import { Institution } from "../../../../../../definitions/services/Institution"
 import { Institutions } from "../../../../../../definitions/services/Institutions";
 import { OrganizationFiscalCode } from "../../../../../../definitions/services/OrganizationFiscalCode";
 import { withRefreshApiCall } from "../../../../authentication/fastLogin/saga/utils";
-import { ServicesClient } from "../../../common/api/__mocks__/client";
+import { ServicesClient } from "../../../common/api/__mocks__/servicesClient";
 import { featuredInstitutionsGet } from "../../store/actions";
 import { handleGetFeaturedInstitutions } from "../handleGetFeaturedInstitutions";
