+++ ts/features/services/search/saga/__tests__/handleFindInstitutions.test.ts
@@ -5,7 +5,7 @@ import { getType } from "typesafe-actions";
 import { InstitutionsResource } from "../../../../../../definitions/services/InstitutionsResource";
 import { OrganizationFiscalCode } from "../../../../../../definitions/services/OrganizationFiscalCode";
 import { withRefreshApiCall } from "../../../../authentication/fastLogin/saga/utils";
-import { ServicesClient } from "../../../common/api/__mocks__/client";
+import { ServicesClient } from "../../../common/api/__mocks__/servicesClient";
 import {
   SearchPaginatedInstitutionsGetPayload,
   searchPaginatedInstitutionsGet
