+++ ts/features/services/search/saga/handleFindInstitutions.ts
@@ -6,7 +6,7 @@ import { SagaCallReturnType } from "../../../../types/utils";
 import { getGenericError, getNetworkError } from "../../../../utils/errors";
 import { readablePrivacyReport } from "../../../../utils/reporters";
 import { withRefreshApiCall } from "../../../authentication/fastLogin/saga/utils";
-import { ServicesClient } from "../../common/api/client";
+import { ServicesClient } from "../../common/api/servicesClient";
 import { searchPaginatedInstitutionsGet } from "../store/actions";
 
 const DEBOUNCE_SEARCH: Millisecond = 500 as Millisecond;
