+++ ts/features/services/search/saga/index.ts
@@ -1,6 +1,6 @@
 import { SagaIterator } from "redux-saga";
 import { call, race, take, takeLatest } from "typed-redux-saga/macro";
-import { ServicesClient } from "../../common/api/client";
+import { ServicesClient } from "../../common/api/servicesClient";
 import { searchPaginatedInstitutionsGet } from "../store/actions";
 import { handleFindInstitutions } from "./handleFindInstitutions";
